Refactor: Update notify to v8.0 and flatten module structure
- Update notify crate from v6.1 to v8.0 (no API changes needed)
- Rename src/ignore.rs to src/file_filter.rs to avoid confusion with ignore crate
- Flatten directory structure by converting single-file modules:
  - src/symbols/mod.rs → src/symbols.rs
  - src/search/mod.rs → src/search.rs
  - src/watcher/mod.rs → src/watcher.rs
