+++ pylight/Cargo.toml
@@ -40,7 +40,7 @@ url = "2.5.4"
 radix_trie = "0.2.1"
 tiny_http = "0.12"
 urlencoding = "2.1"
-notify = "6.1"
+notify = "8.0"
 
 [build-dependencies]
 chrono = "0.4"
