+++ pylight/src/search.rs
