+++ pylight/src/symbols.rs
